Analysis of Chicago_Socioenonomic_data using py
The city of Chicago released a dataset of socioeconomic data to the Chicago City Portal. This dataset contains a selection of six socioeconomic indicators of public health significance and a “hardship index,” for each Chicago community area, for the years 2008 – 2012.  Scores on the hardship index can range from 1 to 100, with a higher index number representing a greater level of hardship. Here we do a basis analysizing of the dataset using Python